% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-generics.R
\name{autocorrelation}
\alias{autocorrelation}
\alias{autocorrelation.default}
\alias{autocorrelation.matrix}
\alias{autocorrelation.FitLandDF}
\title{Autocorrelation Metric for a GLCM}
\usage{
autocorrelation(x, ...)

\method{autocorrelation}{default}(x, ...)

\method{autocorrelation}{matrix}(x, ...)

\method{autocorrelation}{FitLandDF}(x, nlevels, ...)
}
\arguments{
\item{x}{gray-level co-occurrence matrix}

\item{...}{additional parameters}

\item{nlevels}{desired number of discrete gray levels}
}
\value{
autocorrelation metric of \code{x}
}
\description{
Calculate the autocorrelation feature or metric for a gray-level co-occurrence
matrix. For definition and application, see Lofstedt et al. (2019)
\doi{10.1371/journal.pone.0212110}.
}
\examples{
## calculate autocorrelation of arbitrary GLCM
# define arbitrary GLCM
x <- matrix(1:16, nrow = 4)

# normalize
n_x <- normalize_glcm(x)

# calculate autocorrelation
autocorrelation(n_x)

## calculate autocorrelation of arbitrary fitness landscape
# create fitness landscape using FitLandDF object
vals <- runif(64)
vals <- array(vals, dim = rep(4, 3))
my_landscape <- fitscape::FitLandDF(vals)

# calculate autocorrelation of fitness landscape, assuming 2 discrete gray levels
autocorrelation(my_landscape, nlevels = 2)

## confirm value of autocorrelation for fitness landscape
# extract normalized GLCM from fitness landscape
my_glcm <- get_comatrix(my_landscape, discrete = equal_discrete(2))

# calculate autocorrelation of extracted GLCM
autocorrelation(my_glcm)  # should match value of above autocorrelation function call
}
