% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-generics.R
\name{compute_all_metrics}
\alias{compute_all_metrics}
\title{Convenience function to compute all haralick texture features}
\usage{
compute_all_metrics(x)
}
\arguments{
\item{x}{Input glcm matrix}
}
\value{
data.frame with 11 columns
}
\description{
Calculate set of the following Haralick texture features; contrast, entropy, energy,
autocorrelation, correlation, cluster prominence, cluster shade, homogeneity,
inverse difference, max probability, and sum of squares.
}
