% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glcm-marginals.R
\name{differenceEntropy.matrix}
\alias{differenceEntropy.matrix}
\title{Difference entropy is the entropy of marginal distribution of the
difference in gray-level value equivalents x-y}
\usage{
differenceEntropy.matrix(glcm, base = 2)
}
\arguments{
\item{glcm}{gray level co-occurrence matrix}

\item{base}{Base of the logarithm in differenceEntropy.}
}
\value{
float (single value: the entropy of the marginal distribution)
}
\description{
Difference entropy is the entropy of marginal distribution of the
difference in gray-level value equivalents x-y
}
\examples{
# Calculate difference entropy of a given glcm (e.g. uniform matrix)
differenceEntropy.matrix(matrix(1,3,3))

}
