% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{discretize}
\alias{discretize}
\title{Discretize Numeric Variable Into Categories}
\usage{
discretize(x, ...)
}
\arguments{
\item{x}{either a vector (\code{numeric} or \code{integer}) or \code{FitLandDF} object}
}
\value{
discretized form of \code{x}
}
\description{
Takes a numeric variable (could be of class \code{numeric} or \code{integer}) and
returns a discretized version, in which each element has been replaced by a
single integer between \code{1} and \code{nlevels}, inclusive. The \code{nlevels} parameter
is not included in the function definition, but is required and must be a
positive integer value.
}
\examples{

## discretize a numeric vector
vec <- 1:10
discretize(vec, nlevels = 5) # discretize into 5 categories
discretize(vec, 2)           # discretize into 2 categories

## discretize a fitness landscape
# create a 3x3x3 fitness landscape with values 1 through 27
fl_data <- array(1:27, dim = rep(3, 3))
my_fl <- fitscape::FitLandDF(fl_data)
discretize(my_fl, nlevels = 2) # discretize landscape into 2 categories
discretize(my_fl, 5)           # discretize landscape into 5 categories
}
