% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-func.R
\name{get_comatrix}
\alias{get_comatrix}
\title{Calculate Co-Occurrence Matrix}
\usage{
get_comatrix(x, ...)
}
\description{
Calculate generalized co-occurrence matrix from a variety of objects,
currently including fitness landscapes stored as a \code{FitLandDF} instance from
the \code{fitscape} package.
}
