% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glcm-marginals.R
\name{glcm_marginals}
\alias{glcm_marginals}
\alias{xplusy_k}
\alias{xminusy_k}
\title{Marginal distributions of the GLCM}
\usage{
xplusy_k(glcm, k)

xminusy_k(glcm, k)
}
\arguments{
\item{glcm}{square co-occurrence matrix}

\item{k}{real integer (given difference)}
}
\value{
int or double (xplusy_k: sum of matrix entries with given index sum)

int or double (sum of matrix entries with given index difference)
}
\description{
Functions for the calculation of marginal distributions from the GLCM matrix.

The partial sum of the matrix is used to determine the distribution over
the sum of neighbor pairs, returns a value for a given sum k.

The partial sum of the matrix is used to determine the distribution over
the sum of neighbor pairs, returns a value for a given difference of k.
}
