% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glcm-metrics.R
\name{glcm_metrics}
\alias{glcm_metrics}
\alias{xplusy_k}
\alias{glcm_variance}
\title{GLCM Metrics}
\usage{
xplusy_k(glcm, k)

glcm_variance(glcm)
}
\arguments{
\item{glcm}{gray level co-occurrence matrix}

\item{k}{summation equal to k}
}
\description{
GLCM Metrics
}
\section{Functions}{
\itemize{
\item \code{glcm_variance()}: Variance

}}
