% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{quantile_discrete}
\alias{quantile_discrete}
\title{Function to discretize based on quantiles}
\usage{
quantile_discrete(nlevels)
}
\arguments{
\item{nlevels}{number of levels to split continuous vector into}
}
\value{
function that makes a numeric vector discrete
}
\description{
Function to discretize based on quantiles
}
\examples{
# test data
}
