% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-generics.R
\name{compute_all_metrics}
\alias{compute_all_metrics}
\title{Convenience function to compute all haralick texture features for a given comat}
\usage{
compute_all_metrics(x)
}
\arguments{
\item{x}{matrix computed glcm}
}
\value{
data.frame
}
\description{
Convenience function to compute all haralick texture features for a given comat
}
